Les fonctions pour le décalage ont été mises dans un package "model". Le main.py contient la structure principale de notre programme et fait appel à la fonction brute force automatique.
